add @diagnostics field - sf id: 1887564, re #1
This patch adds a new field, called @diagnostics, which is treated like
other indexed fields, such as @todo, and @bug.

The idea behind the patch is to allow developers to create diagnostics
sections in their documentation, that contains a list of error messages
that the code generates, along with a snippet of text about what it
means. For example:

@diagnostics: C{Unrecognized token in config file: %s}

The config file contained an invalid token. This may be caused by a
typo, if you edited the file by hand, or the pluralization of a keyword
where the singular was expected. If you haven't changed the file by
hand, the config file may have been created with an earlier version of
the software, which had a slightly different format. See the "Config
Files" section if the upgrade documentation.

The attached patch fixes the epydoc docstring parser, docwriter, and the
documentation. It makes no changes to documentation generated in Latex,
or with the javaDoc or reStructuredText formats.

Testing has been adhoc, and no additional test cases were added to the
doctests.

https://sourceforge.net/support/tracker.php?aid=1887564

Signed-off-by: Johannes Dewender <github@JonnyJD.net>