allow from .foo import * (relative) sf 3309660
patch from
https://sourceforge.net/support/tracker.php?aid=3309668

Tested on 3.0.1

A relative import like:
from .foo import *

is not parsed correctly.

| Source code parsing failed (but introspection was successful).
| Error: Error during parsing: invalid syntax
| -- Bad dotted name

With the attached patch it does parse correctly.

----

Note that general support for relative imports was added with
d38a5778a0817b0b968c5c5b632ece7fa7ccfad8 in 2007-03-10.
This is just a special case, that was ommited before.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>